% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_percentile_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{mean_ci function - calculates the mean and uses non-parametric bootstrapping
to calculate the confidence intervals}
\usage{
mean_ci(observations, bootstraps = 1e+05, conf = 0.95, type = "bca")
}
\description{
mean_ci function - calculates the mean and uses non-parametric bootstrapping
to calculate the confidence intervals
}
