% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_percentile.R
\name{weib_percentile}
\alias{weib_percentile}
\title{Calculating a percentile estimate of a seasonal abundance distribution from
incidental observations.}
\usage{
weib_percentile(observations, percentile = 0.9, iterations = 100)
}
\arguments{
\item{observations}{is a vector of dates/time of observations given as integers}

\item{percentile}{is the percentile of the cumulative distribution function of interest}

\item{iterations}{is the number of iterations you want to run to create empirical
bootstrapping to estimate bias of original CDF. The bias is used to calculate
a bias corrected estimate of the percentile bound.}
}
\description{
\code{weib_percentile} uses empirical bootstrapping to estimate a percentile
of the Weibull distribution, given random variables.
}
\examples{

Gather 10 sightings of an individuals - numbers represent day of year observed
testobs <- c(150,160,162,164,168,170,172,176,178,188)

Estimate when 90\% of individuals are still in the phenological state (e.g.,
when 90\% of plants are still in flower or when 90\% of butterflies are still on wing)
weib_percentile(observations, percentile = 0.9)

Estimate when 50\% of ndividuals are still in the phenological state
with 50 iterations for quicker processing
weib_percentile(observations, percentile = 0.5, iterations = 50)
}
\keyword{percentile}
\keyword{phenology,}
\keyword{weibull,}
