% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_percentile_ci.R
\name{quantile_ci}
\alias{quantile_ci}
\title{Calculating the CIs of a naive (non-parameterized) estimate of a percentile of
a vector of observations using non-parametric bootstrapping}
\usage{
quantile_ci(observations, percentile, bootstraps = 1e+05, conf = 0.95,
  type = "bca")
}
\arguments{
\item{observations}{A vector of observations given as numeric values}

\item{percentile}{The naive percentile of interest}

\item{bootstraps}{The number of bootstraps you want to run to create the CIs,
defaults to 100000}

\item{conf}{The confidence level wanted. Defaults to 95% CI.}

\item{type}{A vector of character strings represenging the type of intervals
required to calculate the CI. Defaults to "bca". See ??boot.ci for more information.}
}
\description{
\code{naive_percentile} Estimates CI around a naive percentile estimate using
non-parameteric bootstrapping from the boot package
}
\examples{

#' # Gather last 10 sightings of dodo
observations <- c(1662, 1638, 1631, 1628, 1628, 1611, 1607, 1602, 1601, 1598)

# Estimate Naive percentile when 90\% of dodo were left with 95\% CI given last observations
weib_percentile_ci(observations, iterations = 100, percentile = 0.9, bootstraps = 100)
}
