% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R, R/naive_percentile_ci.R,
%   R/quantile_ci.R
\name{mean_ci}
\alias{mean_ci}
\alias{quantile_ci}
\title{Calculating the confidence intervals (CIs) of a mean estimate of a
a vector of observations using non-parametric bootstrapping.}
\usage{
mean_ci(observations, bootstraps = 1e+05, conf = 0.95, type = "bca")

quantile_ci(observations, percentile, bootstraps = 1e+05, conf = 0.95,
  type = "bca")

mean_ci(observations, bootstraps = 1e+05, conf = 0.95, type = "bca")

quantile_ci(observations, percentile, bootstraps = 1e+05, conf = 0.95,
  type = "bca")
}
\arguments{
\item{observations}{A vector of observations given as numeric values}

\item{bootstraps}{The number of bootstraps you want to run to create the CIs,
defaults to 100000}

\item{conf}{The confidence level wanted. Defaults to 95\% CI.}

\item{type}{A vector of character strings represenging the type of intervals
required to calculate the CI. Defaults to "bca". See ??boot.ci for more information.}

\item{percentile}{The percentile of interest}

\item{observations}{A vector of observations given as numeric values}

\item{percentile}{The percentile of interest}

\item{bootstraps}{The number of bootstraps you want to run to create the CIs,
defaults to 100000}

\item{conf}{The confidence level wanted. Defaults to 95\% CI.}

\item{type}{A vector of character strings represenging the type of intervals
required to calculate the CI. Defaults to "bca". See ??boot.ci for more information.}

\item{observations}{A vector of observations given as numeric values}

\item{percentile}{The percentile of interest}

\item{bootstraps}{The number of bootstraps you want to run to create the CIs,
defaults to 100000}

\item{conf}{The confidence level wanted. Defaults to 95\% CI.}

\item{type}{A vector of character strings represenging the type of intervals
required to calculate the CI. Defaults to "bca". See ??boot.ci for more information.}
}
\description{
Functions that estimate CIs using nonparametric bootstrapping around a mean estimate.

#' \code{mean_ci} Estimates CIs around a mean estimate using non-parametric bootstrapping
from the boot package

Functions that estimate CIs using nonparametric bootstrapping.

\code{quantile_ci} Estimates CIs around a quantile percentile estimate using
non-parameteric bootstrapping from the boot package

\code{mean_ci} Estimates CIs around a mean estimate using non-parametric bootstrapping
from the boot package

Functions that estimate CIs using nonparametric bootstrapping around quantile estimates.

\code{quantile_ci} Estimates CIs around a quantile percentile estimate using
non-parameteric bootstrapping from the boot package
}
\section{Functions}{
\itemize{
\item \code{mean_ci}: Estimates CIs around a quantile percentile estimate using
non-parameteric bootstrapping from the boot package

\item \code{mean_ci}: Estimates CIs around a quantile percentile estimate using
non-parameteric bootstrapping from the boot package
}}

\keyword{mean}
\keyword{percentile}
\keyword{phenology}
\keyword{quantile}
