% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_percentile.R
\name{weib_percentile}
\alias{weib_percentile}
\title{Calculating the percentile bound of a distribution of onset/cessation.}
\usage{
weib_percentile(observations, percentile = 0.9, iterations = 100)
}
\arguments{
\item{observations}{is a vector of dates/time of observations given as integers}

\item{percentile}{is the percentile of the cumulative distribution function of interest}

\item{iterations}{is the number of iterations you want to run to create empirical
bootstrapping to estimate bias of original CDF. The bias is used to calculate
a bias corrected estimate of the percentile bound.}
}
\description{
\code{weib_percentile} uses empirical bootstrapping to estimate a percentile
of the Weibull distribution, given random variables.
}
\examples{

# Gather last 10 sightings of dodo
observations <- c(1662, 1638, 1631, 1628, 1628, 1611, 1607, 1602, 1601, 1598)

# Estimate when 90\% of dodo were left
weib_percentile(observations, percentile = 0.9)

Estimate when 50\% of dodo were left with 50 iterations for quicker processing
weib_percentile(observations, percentile = 0.5, iterations = 50)
}
