% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_percentile_ci.R
\name{weib_percentile_ci}
\alias{weib_percentile_ci}
\title{Calculating the CIs of a percentile estimate of a seasonal abundance distribution
using the non-parametric bootstrapping.}
\usage{
weib_percentile_ci(observations, iterations, percentile, bootstraps)
}
\arguments{
\item{observations}{is a vector of dates/time of observations given as numeric values}

\item{iterations}{is the number of iterations you want to run to create empirical
bootstrapping to estimate bias of original CDF. The bias is used to calculate
a bias corrected estimate of the percentile bound.}

\item{percentile}{is the percentile of the cumulative distribution function of interest}

\item{bootstraps}{is the number of bootstraps you want to run to create the CIs}
}
\description{
\code{weib_percentile_ci} uses non-parametric bootstrapping from the boot package
to estimate 95% CIs
}
\examples{

Gather 10 sightings of an individuals - numbers represent day of year observed
testobs <- c(150,160,162,164,168,170,172,176,178,188)

Estimate when 90\% of individuals are still in the phenological state (e.g.,
when 90\% of plants are still in flower or when 90\% of butterflies are still on wing)
weib_percentile_ci(observations, iterations = 100, percentile = 0.9, bootstraps = 100)
}
\keyword{percentile}
\keyword{phenology,}
\keyword{weibull,}
